backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
argon2-cffi==23.1.0
itsdangerous==2.2.0
httpx==0.27.2
google-generativeai==0.8.3
